Fix: incorrect types (#212)
* Revert "Revert "Fix imotioncube types to resemble the CoE manual""

This reverts commit b048b318b4e2a73cb4ad0e887de0f011bb774591.

* Fix pdo map base_sync_manager type

* Fix types in joint

* Fix types used in ethercat

* Remove unused angle conversions

* Change radians to from float to double

the controller uses double and there is no clear advantage in using float

* Fix types in hardware interface

* fix style

* Fix status word type

* Fix some more types in hardware interface

+++ march_hardware/include/march_hardware/ActuationMode.h
@@ -46,7 +46,7 @@ public:
     }
   }
 
-  int toModeNumber()
+  uint8_t toModeNumber()
   {
     if (this->value_ == position)
     {
