+++ march_hardware/include/march_hardware/Encoder.h
@@ -13,11 +13,11 @@ private:
   int slaveIndex;
   int totalPositions;
 
-  int upperHardLimitIU;
-  int lowerHardLimitIU;
-  int upperSoftLimitIU;
-  int lowerSoftLimitIU;
-  int zeroPositionIU;
+  int32_t upperHardLimitIU;
+  int32_t lowerHardLimitIU;
+  int32_t upperSoftLimitIU;
+  int32_t lowerSoftLimitIU;
+  int32_t zeroPositionIU;
 
   float safetyMarginRad;
 
@@ -33,26 +33,27 @@ public:
     , safetyMarginRad(0)
   {
   }
-  Encoder(int numberOfBits, int minPositionIU, int maxPositionIU, int zeroPositionIU, float safetyMarginRad);
+  Encoder(int numberOfBits, int32_t minPositionIU, int32_t maxPositionIU, int32_t zeroPositionIU,
+          float safetyMarginRad);
 
-  float getAngleRad(uint8_t ActualPositionByteOffset);
+  double getAngleRad(uint8_t ActualPositionByteOffset);
 
-  int getAngleIU(uint8_t ActualPositionByteOffset);
+  int32_t getAngleIU(uint8_t ActualPositionByteOffset);
 
-  float IUtoRad(int iu);
-  int RadtoIU(float rad);
+  double IUtoRad(int32_t iu);
+  int32_t RadtoIU(double rad);
 
-  bool isWithinHardLimitsIU(int positionIU);
-  bool isWithinSoftLimitsIU(int positionIU);
-  bool isValidTargetIU(int currentIU, int targetIU);
+  bool isWithinHardLimitsIU(int32_t positionIU);
+  bool isWithinSoftLimitsIU(int32_t positionIU);
+  bool isValidTargetIU(int32_t currentIU, int32_t targetIU);
 
   void setSlaveIndex(int slaveIndex);
 
   int getSlaveIndex() const;
-  int getUpperSoftLimitIU() const;
-  int getLowerSoftLimitIU() const;
-  int getUpperHardLimitIU() const;
-  int getLowerHardLimitIU() const;
+  int32_t getUpperSoftLimitIU() const;
+  int32_t getLowerSoftLimitIU() const;
+  int32_t getUpperHardLimitIU() const;
+  int32_t getLowerHardLimitIU() const;
 
   /** @brief Override comparison operator */
   friend bool operator==(const Encoder& lhs, const Encoder& rhs)
